requirements.txt
flask
flask_sqlalchemy
python-docx
scikit-learn
pandas
joblib
mysql-connector-python
python-dotenv
reportlab